Augmentation de la taille des paramètres (uint8 -> uint32)

+++ inc/utils.h
@@ -9,8 +9,8 @@
     } while(0)
 
 /* Fonctions pour tableaux */
-byte *init_array(uint8 length);
-byte **init_2d_array(uint8 length1, uint8 length2);
+byte *init_array(uint32 length);
+byte **init_2d_array(uint32 length1, uint32 length2);
 void expand_array(byte **array, uint32 new_length);
 void free_array(byte **array, uint32 length);
 void free_2d_array(byte ***array, uint32 length);
